+++ src/apps/settings/modules/information/infrastructure.tsx
@@ -1,13 +1,18 @@
 import { Icon } from '../../../utils/components/Icon';
 import { SettingsGroup, SettingsOption, SettingsSubGroup } from '../../components/SettingsBox';
+import { invoke } from '@tauri-apps/api/core';
 import { exit, relaunch } from '@tauri-apps/plugin-process';
 import { Button } from 'antd';
 
-import { EnvConfig } from '../shared/config/infra';
+import { EnvConfig, getSettingsPath } from '../shared/config/infra';
 import { LoadCustomConfigFile } from './infra.api';
 import cs from './infra.module.css';
 
 export function Information() {
+  const openSettingsFile = async () => {
+    invoke('open_file', { path: await getSettingsPath('settings.json') });
+  };
+
   return (
     <div className={cs.info}>
       <SettingsGroup>
@@ -36,7 +41,11 @@ export function Information() {
 
       <SettingsGroup>
         <SettingsOption>
-          <span>Load config file (will replace current configurations):</span>
+          <span>Open Settings file</span>
+          <Button onClick={openSettingsFile}>Open</Button>
+        </SettingsOption>
+        <SettingsOption>
+          <span>Load Custom file (will replace current settings):</span>
           <Button onClick={LoadCustomConfigFile}>Select File</Button>
         </SettingsOption>
       </SettingsGroup>
@@ -44,13 +53,13 @@ export function Information() {
       <SettingsGroup>
         <SettingsOption>
           <span>Force Restart</span>
-          <Button type="dashed" onClick={relaunch} style={{ width: '40px' }}>
+          <Button type="dashed" onClick={relaunch} style={{ width: '50px' }}>
             <Icon iconName="IoReload" propsIcon={{ size: 12 }} />
           </Button>
         </SettingsOption>
         <SettingsOption>
-          <span>Quit/Close</span>
-          <Button type="dashed" onClick={() => exit(0)} style={{ width: '40px' }}>
+          <span>Quit/Close Seelen UI</span>
+          <Button type="dashed" danger onClick={() => exit(0)} style={{ width: '50px' }}>
             <Icon iconName="IoClose" />
           </Button>
         </SettingsOption>