+++ src/apps/settings/modules/shared/config/infra.ts
@@ -1,3 +1,9 @@
+import { path } from '@tauri-apps/api';
+
 export const EnvConfig = {
   version: process.env.packageVersion,
 };
+
+export async function getSettingsPath(...sub: string[]) {
+  return await path.join(await path.homeDir(), '.config', 'seelen', ...sub);
+}