+++ src/apps/settings/modules/shared/store/storeApi.ts
@@ -8,6 +8,7 @@ import { path } from '@tauri-apps/api';
 import { convertFileSrc, invoke } from '@tauri-apps/api/core';
 import yaml from 'js-yaml';
 
+import { getSettingsPath } from '../config/infra';
 import { dialog, fs } from '../tauri/infra';
 
 import { AppsTemplates } from '../../../../utils/appsTemplates';
@@ -141,9 +142,8 @@ export async function loadUserSettings(route?: string): Promise<UserSettings> {
     env: await invoke('get_user_envs'),
   };
 
-  const json_route =
-    route || (await path.join(await path.homeDir(), '.config/seelen/settings.json'));
-  const yaml_route = await path.join(await path.homeDir(), '.config/seelen/applications.yml');
+  const json_route = route || (await getSettingsPath('settings.json'));
+  const yaml_route = await getSettingsPath('applications.yml');
 
   if (await fs.exists(json_route)) {
     userSettings.jsonSettings = parseAsCamel(
@@ -194,14 +194,17 @@ export async function createAhkFiles(ahkVariables: AhkVariables) {
       content = content.replace(/{{(.*?)}}/g, (match, varname) => {
         return ahkVariables[varname]?.ahk || match;
       });
-      await fs.writeTextFile(await path.join(staticPath, entry.name.replace('.template', '')), content);
+      await fs.writeTextFile(
+        await path.join(staticPath, entry.name.replace('.template', '')),
+        content,
+      );
     }
   }
 }
 
 export async function saveUserSettings(settings: UserSettings) {
-  const json_route = await path.join(await path.homeDir(), '.config/seelen/settings.json');
-  const yaml_route = await path.join(await path.homeDir(), '.config/seelen/applications.yml');
+  const json_route = await getSettingsPath('settings.json');
+  const yaml_route = await getSettingsPath('applications.yml');
 
   if (settings.jsonSettings.ahkEnabled) {
     await createAhkFiles(settings.jsonSettings.ahkVariables);